Move MainAction to new picard.ui.enums
- prevent circular imports
- no need to initialize mainwindow module when only imported this enum
- move mainwindow/create.py to mainwindow/actions.py (it makes more sense)
- drop local imports that were added to prevent circular imports
