+++ picard/cluster.py
@@ -65,6 +65,7 @@ from picard.util.imagelist import (
     update_metadata_images,
 )
 
+from picard.ui.enums import MainAction
 from picard.ui.item import (
     FileListItem,
     Item,
@@ -337,12 +338,10 @@ class UnclusteredFiles(Cluster):
 
     def add_files(self, files, new_album=True):
         super().add_files(files, new_album=new_album)
-        from picard.ui.mainwindow.actions import MainAction
         self.tagger.window.enable_action(MainAction.CLUSTER, self.files)
 
     def remove_file(self, file, new_album=True):
         super().remove_file(file, new_album=new_album)
-        from picard.ui.mainwindow.actions import MainAction
         self.tagger.window.enable_action(MainAction.CLUSTER, self.files)
 
     def lookup_metadata(self):
